% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{add_index_user}
\alias{add_index_user}
\title{Add index user}
\usage{
add_index_user(index, email, role, credentials = NULL)
}
\arguments{
\item{index}{The index to list}

\item{email}{The email of an (existing) user}

\item{role}{The role of the user (METAREADER, READER, WRITER, ADMIN)}
}
\description{
Add index user
}
