% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{amcat_auth}
\alias{amcat_auth}
\title{Authenticate to an AmCAT instance}
\usage{
amcat_auth(server, refresh = FALSE)
}
\arguments{
\item{server}{URL of the AmCAT instance}

\item{refresh}{Whether to enable refresh token rotation (see details).}
}
\value{
an amcat4_token object
}
\description{
Authenticate to an AmCAT instance
}
\details{
Enabling refresh token rotation ensures added security as leaked
  refresh tokens also become invalidated after a short while. It is currently
  disabled by default as it is not fully supported by the underlying httr2
  package.

  The function needs to open a browser, which will usually only work in an
  interactive session. However, you can save the returned object in an rds
  file (with \code{saveRDS()}) and tell amcat4r where to look for it:
  \code{options(amcat4r_token_cache = "path/to/location/tokens.rds")}. If you
  still have issues in an interactive session, check \link[utils]{browseURL}
  to see if you can set a browser manually.
}
\examples{
\dontrun{
  amcat_auth("https://middlecat.up.railway.app/api/demo_resource")
}
}
