% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{create_user}
\alias{create_user}
\title{Create a new user}
\usage{
create_user(email, role = "writer", index_access = NULL, credentials = NULL)
}
\arguments{
\item{email}{email of the user to add.}

\item{role}{global role of the user ("metareader", "reader", "writer" or
"admin").}

\item{index_access}{index to grant access to for the new user.}

\item{credentials}{The credentials to use. If not given, uses cached login
information.}
}
\description{
Create a new user
}
