% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{delete_by_query}
\alias{delete_by_query}
\title{Delete documents by query}
\usage{
delete_by_query(
  index,
  ids = NULL,
  queries = NULL,
  filters = NULL,
  credentials = NULL
)
}
\arguments{
\item{index}{The index to query}

\item{ids}{A optional vector of ids to add/remove tags from}

\item{queries}{An optional vector of queries to run (implicit OR)}

\item{filters}{An optional list of filters, e.g. list(publisher='A', date=list(gte='2022-01-01'))}
}
\description{
Delete documents by query
}
\examples{
\dontrun{
 delete_by_query("my_index", filters=list(publisher='NY Times'))
 delete_by_query("my_index", ids=c(42, 69))
 delete_by_query("my_index", queries="advertisement")
}
}
