% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{delete_documents}
\alias{delete_documents}
\title{Delete documents from index}
\usage{
delete_documents(index, docid, credentials = NULL)
}
\arguments{
\item{index}{The index name in which documents should be deleted.}

\item{docid}{the .ids of the documents that should be deleted.}

\item{credentials}{The credentials to use. If not given, uses last login
information.}
}
\description{
Delete documents from index
}
