% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{delete_index}
\alias{delete_index}
\title{Delete an index}
\usage{
delete_index(index, credentials = NULL)
}
\arguments{
\item{credentials}{The credentials to use. If not given, use last login information}
}
\description{
Delete an index
}
