% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{delete_index_user}
\alias{delete_index_user}
\title{Delete index user}
\usage{
delete_index_user(index, email, credentials = NULL)
}
\arguments{
\item{index}{The index to list}

\item{email}{The email of an (existing) user}
}
\description{
Delete index user
}
