% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{delete_user}
\alias{delete_user}
\title{Delete new user}
\usage{
delete_user(email, credentials = NULL)
}
\arguments{
\item{email}{email of the user to remove.}

\item{credentials}{The credentials to use. If not given, uses cached login
information.}
}
\description{
Delete new user
}
