% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{do_delete}
\alias{do_delete}
\title{Execute a DELETE request to this AmCAT API, returning the json result}
\usage{
do_delete(credentials, url, ...)
}
\description{
Execute a DELETE request to this AmCAT API, returning the json result
}
