% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{do_get}
\alias{do_get}
\title{Execute a GET request to this AmCAT API, returning the json result}
\usage{
do_get(credentials, url, ...)
}
\description{
Execute a GET request to this AmCAT API, returning the json result
}
