% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{do_post}
\alias{do_post}
\title{Execute a POST request to this AmCAT API, returning the json result}
\usage{
do_post(credentials, url, body, encode = "json-auto", auto_unbox = TRUE, ...)
}
\description{
Execute a POST request to this AmCAT API, returning the json result
}
