% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{do_put}
\alias{do_put}
\title{Execute a PUT request to this AmCAT API, returning the json result}
\usage{
do_put(credentials, url, body, encode = "json", ...)
}
\description{
Execute a PUT request to this AmCAT API, returning the json result
}
