% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingjobs.R
\name{get_codingjob}
\alias{get_codingjob}
\title{Retrieve a codingjob (including results)}
\usage{
get_codingjob(id, credentials = NULL)
}
\arguments{
\item{id}{the job ID}

\item{credentials}{The credentials to use. If not given, use last login information}
}
\description{
Retrieve a codingjob (including results)
}
