% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{get_credentials}
\alias{get_credentials}
\title{Helper function to get credentials from artugment or pkg.env}
\usage{
get_credentials(credentials = NULL)
}
\description{
Helper function to get credentials from artugment or pkg.env
}
