% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_document}
\alias{get_document}
\title{Retrieve a single document}
\usage{
get_document(index, doc_id, fields, credentials = NULL)
}
\arguments{
\item{index}{The index to get fields for}

\item{doc_id}{A single document_id}

\item{fields}{Optional character vector listing the fields to retrieve}

\item{credentials}{The credentials to use. If not given, uses last login information}
}
\value{
A tibble with one row containing the requested fields
}
\description{
Retrieve a single document
}
