% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_documents}
\alias{get_documents}
\title{Retrieve multiple documents using a purrr map over get_document}
\usage{
get_documents(index, doc_ids, fields, credentials = NULL, ...)
}
\arguments{
\item{index}{The index to get fields for}

\item{doc_ids}{A vector of document_ids}

\item{fields}{Optional character vector listing the fields to retrieve}

\item{credentials}{The credentials to use. If not given, uses last login information}

\item{...Other}{options to pass to map, e.g. .progress}
}
\value{
A tibble with one row containing the requested fields
}
\description{
Retrieve multiple documents using a purrr map over get_document
}
