% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_fields}
\alias{get_fields}
\title{Get fields}
\usage{
get_fields(index, credentials = NULL)
}
\arguments{
\item{index}{The index to get fields for}

\item{credentials}{The credentials to use. If not given, uses last login information}
}
\description{
Get fields
}
