% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_index}
\alias{get_index}
\title{Get a single index}
\usage{
get_index(index, credentials = NULL)
}
\arguments{
\item{index}{name of the index}

\item{credentials}{The credentials to use. If not given, uses last login
information.}
}
\value{
a list with details about this index, or NULL if it does not exist
}
\description{
Get a single index
}
\examples{
\dontrun{
get_index("my_index")
}
}
