% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user}
\alias{get_user}
\title{Get information about a user}
\usage{
get_user(user = "me", credentials = NULL)
}
\arguments{
\item{user}{The user to get information on, or 'me' to get information on the current user}

\item{credentials}{The credentials to use. If not given, uses cached login
information.}
}
\description{
Get information about a user
}
