% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_exists}
\alias{index_exists}
\title{Check if an index exists}
\usage{
index_exists(index, credentials = NULL)
}
\arguments{
\item{index}{name of the index}

\item{credentials}{The credentials to use. If not given, uses last login
information.}
}
\value{
a list with details about this index, or NULL if it does not exist
}
\description{
Convenience function that calls !is.null(get_index(...))
}
\examples{
\dontrun{
index_exists("my_index")
}
}
