% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{list_indexes}
\alias{list_indexes}
\title{List the indexes on this server}
\usage{
list_indexes(credentials = NULL)
}
\arguments{
\item{credentials}{The credentials to use. If not given, uses last login
information.}
}
\value{
a tibble with index names and guest_role
}
\description{
List the indexes on this server
}
\examples{
\dontrun{
list_indexes()
}
}
