% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{list_users}
\alias{list_users}
\title{List users}
\usage{
list_users(credentials = NULL)
}
\arguments{
\item{credentials}{The credentials to use. If not given, uses last login information}
}
\description{
List users
}
