% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{login}
\alias{login}
\title{Logon to AmCAT}
\usage{
login(host, username, password = NULL)
}
\arguments{
\item{host}{The host including schema i.e. "http://localhost:5000"}

\item{username}{The username to login with}

\item{password}{The password to login with. If not given, will prompt interactively}
}
\value{
A named list containing host, username, and token
}
\description{
Logon to AmCAT
}
