% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{modify_index_user}
\alias{modify_index_user}
\title{Modify index user}
\usage{
modify_index_user(index, email, role, credentials = NULL)
}
\arguments{
\item{index}{The index to list}

\item{email}{The email of an (existing) user}

\item{role}{The role of the user (METAREADER, READER, WRITER, ADMIN)}

\item{credentials}{The credentials to use. If not given, uses last login information}
}
\description{
Modify index user
}
