% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{ping}
\alias{ping}
\title{Check if an amcat instance is reachable}
\usage{
ping(server)
}
\arguments{
\item{server}{A character string of the server URL. If missing the server for
the logged in session is tried.}
}
\value{
A logical value indicating if the server is reachable.
}
\description{
Check if a server is reachable by sending a request to its
config endpoint.
}
\examples{
\dontrun{
ping("http://localhost/amcat")
}
}
