% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{post_reindex}
\alias{post_reindex}
\title{Post a reindex request to the API}
\usage{
post_reindex(
  index,
  destination,
  field_options = NULL,
  queries = NULL,
  filters = NULL,
  credentials = NULL
)
}
\arguments{
\item{index}{The source index name}

\item{destination}{The destination index name}

\item{field_options}{Optional named list of per-field options (rename/exclude/type)}

\item{queries}{Optional list of query strings to filter documents}

\item{filters}{Optional list of filters to apply during reindex}

\item{credentials}{The credentials to use. If not given, uses last login information}
}
\description{
Low-level wrapper around the reindex endpoint. For most use cases, use
\code{\link{reindex}} instead.
}
\keyword{internal}
