% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_aggregate}
\alias{query_aggregate}
\title{Conduct a query and return the resulting documents}
\usage{
query_aggregate(
  index,
  axes,
  queries = NULL,
  filters = NULL,
  credentials = NULL
)
}
\arguments{
\item{index}{The index to query}

\item{axes}{The aggregation axes, e.g. list(list(field="publisher", list(field="date", interval="year")))}

\item{queries}{An optional vector of queries to run (implicit OR)}

\item{filters}{An optional list of filters, e.g. list(publisher='A', date=list(gte='2022-01-01'))}

\item{credentials}{The credentials to use. If not given, use last login information}
}
\description{
Conduct a query and return the resulting documents
}
