% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{refresh_index}
\alias{refresh_index}
\title{Refresh an index}
\usage{
refresh_index(index, credentials = NULL)
}
\arguments{
\item{index}{The index to refresh}

\item{credentials}{The credentials to use. If not given, uses last login information}
}
\description{
Refresh an index
}
