% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{request}
\alias{request}
\title{Helper function to execute a request to this API}
\usage{
request(
  credentials,
  url,
  request_function = httr::GET,
  error_on_404 = TRUE,
  ...
)
}
\description{
Helper function to execute a request to this API
}
