% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.r
\name{run_amcat_docker}
\alias{run_amcat_docker}
\title{Run docker containers with AmCAT modules}
\usage{
run_amcat_docker(compose = NULL, force_install = FALSE)
}
\arguments{
\item{compose}{Path to a Docker Compose file. Uses
https://github.com/JBGruber/amcat4docker/blob/main/docker-compose.yml by
default.}

\item{force_install}{If TRUE, removes all containers and re-creates them from
the compose file. If 2, the images are also re-downloaded. Danger: this
will destroy the indexes in your containers!}
}
\description{
Automatically sets up services defined in a Docker Compose file. Only options
relevant to AmCAT are implemented.
}
