% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{set_fields}
\alias{set_fields}
\title{Set fields}
\usage{
set_fields(index, fields, credentials = NULL)
}
\arguments{
\item{index}{The index to set fields for}

\item{fields}{A list with fields and data types, e.g. list(author="keyword")}
}
\description{
Set fields
}
