% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.r
\name{stop_amcat_docker}
\alias{stop_amcat_docker}
\title{Stop docker containers with AmCAT modules}
\usage{
stop_amcat_docker(filters = c("elastic", "amcat"))
}
\arguments{
\item{filters}{Names of containers or named values for other filters. (See
[https://docs.docker.com/engine/api/v1.41/#tag/Container/operation/ContainerList])}
}
\description{
Stop docker containers with AmCAT modules
}
\examples{
\dontrun{
# stop AmCAT modules
stop_amcat_docker()

# stop container by id
stop_amcat_docker(filters = c(id = "a6cbe4787227"))

# stop all containers
stop_amcat_docker(filters = "")
}
}
