% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_old.R
\name{test_credentials}
\alias{test_credentials}
\title{Test amcat4r credentials object}
\usage{
test_credentials(credentials)
}
\arguments{
\item{credentials}{amcat4r credentials object}
}
\description{
Test amcat4r credentials object as returned by \code{login()}
}
