% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{update_by_query}
\alias{update_by_query}
\title{Update documents by query}
\usage{
update_by_query(
  index,
  field,
  value,
  ids = NULL,
  queries = NULL,
  filters = NULL,
  credentials = NULL
)
}
\arguments{
\item{index}{The index to query}

\item{field}{The field name to update}

\item{value}{The new value for the field}

\item{ids}{A optional vector of ids to add/remove tags from}

\item{queries}{An optional vector of queries to run (implicit OR)}

\item{filters}{An optional list of filters, e.g. list(publisher='A', date=list(gte='2022-01-01'))}
}
\description{
Update documents by query
}
\examples{
\dontrun{
   update_by_query("my_index", "publisher", "NYT", filters=list(publisher='New York Times'))
   update_by_query("my_index", "sentiment", -1, ids=c(3, 7, 9, 11))
}
}
