% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{update_documents}
\alias{update_documents}
\title{Update documents}
\usage{
update_documents(index, ids = NULL, documents, credentials = NULL)
}
\arguments{
\item{index}{The index name to create.}

\item{ids}{The IDs (.id) of the document to update description (if NULL, the
.id column from documents will be used).}

\item{documents}{A data frame with columns to update.}

\item{credentials}{The credentials to use. If not given, uses last login
information.}
}
\description{
Update documents
}
