% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{update_tags}
\alias{update_tags}
\title{Add or remove tags to/from documents by query or ID}
\usage{
update_tags(
  index,
  action,
  field,
  tag,
  ids = NULL,
  queries = NULL,
  filters = NULL,
  credentials = NULL
)
}
\arguments{
\item{index}{The index to query}

\item{action}{'add' or 'remove' the tags}

\item{field}{The tag field name}

\item{tag}{The tag to add or remove}

\item{ids}{A vector of ids to add/remove tags from}

\item{queries}{An optional vector of queries to run (implicit OR)}

\item{filters}{An optional list of filters, e.g. list(publisher='A', date=list(gte='2022-01-01'))}

\item{credentials}{The credentials to use. If not given, use last login information}
}
\description{
Add or remove tags to/from documents by query or ID
}
