% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.R
\name{create_index}
\alias{create_index}
\title{Create an index}
\usage{
create_index(index, credentials = NULL)
}
\arguments{
\item{index}{Name of the index to create}

\item{credentials}{The credentials to use. If not given, use last login information}
}
\description{
Create an index
}
