% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{modify_user}
\alias{modify_user}
\title{Modify an existing user}
\usage{
modify_user(email, role = "writer", new_password = NULL, credentials = NULL)
}
\arguments{
\item{email}{email of the user to modify.}

\item{role}{global role of the user ("metareader", "reader", "writer" or
"admin").}

\item{new_password}{new password for the user.}

\item{credentials}{The credentials to use. If not given, uses cached login
information.}
}
\description{
Modify an existing user
}
