% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amcatr.R
\name{post_codingjob}
\alias{post_codingjob}
\title{Post a codingjob to the annotation backend}
\usage{
post_codingjob(codingjob, credentials = NULL)
}
\arguments{
\item{codingjob}{The job as a (json) list}

\item{credentials}{The credentials to use. If not given, use last login information}
}
\description{
Post a codingjob to the annotation backend
}
