% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.r
\name{run_amcat_docker}
\alias{run_amcat_docker}
\title{Run docker containers with AmCAT modules}
\usage{
run_amcat_docker(compose = NULL, force = FALSE)
}
\arguments{
\item{compose}{Path to a Docker Compose file. Uses
https://github.com/JBGruber/amcat4docker/blob/main/docker-compose.yml by
default.}

\item{force}{If TRUE, removes all containers and images mentioned in the
compose file and re-downloads them.}
}
\description{
Automatically sets up services defined in a Docker Compose file. Only options
relevant to AmCAT are implemented.
}
