% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{set_fields}
\alias{set_fields}
\title{Set fields}
\usage{
set_fields(index, fields, credentials = NULL)
}
\arguments{
\item{index}{The index to set fields for}

\item{fields}{A list with fields and data types, e.g. list(author="keyword")}

\item{credentials}{The credentials to use. If not given, uses last login information}
}
\description{
Set fields
}
\details{
AmCAT currently supports the following field types:
\itemize{
\item \strong{text}: For general text columns
\item \strong{keyword}: Keywords - this is like text, but is not parsed as words. Most suitable for 'factor' / 'group' level data
\item \strong{tag}: Tags - like keywords, but the assumption is that every document can have multiple values.
\item \strong{date}: For date or date+time columns
\item \strong{boolean}: Boolean (true/false) columns
\item \strong{number}: General numberic columns
\item \strong{integer}: Whole numbers
\item \strong{object}: Nested dictionaries. These are not really analysed by AmCAT, but can store any data you need
\item \strong{json}: Generic json data. This is analysed (and searchable) as text, use object if AmCAT does not need to search it
\item \strong{vector}: Dense vectors, useful for e.g. embedding vectors
\item \strong{geo_point}: Geometrical locations (long+lat)
\item \strong{url}: A generic URL
\item \strong{image}, \strong{video}: A URL pointing to an image or video file
}
}
