% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.r
\name{stop_amcat_docker}
\alias{stop_amcat_docker}
\title{Stop docker containers with AmCAT modules}
\usage{
stop_amcat_docker(compose = NULL, filters = NULL)
}
\arguments{
\item{compose}{Path to a Docker Compose file. Uses
https://github.com/JBGruber/amcat4docker/blob/main/docker-compose.yml by
default.}

\item{filters}{Names of containers or named values for other filters.}
}
\description{
Stop docker containers with AmCAT modules
}
\details{
Stops either the containers defined in a compose file or the
filters. If filters is set, compose is ignored.
}
\examples{
\dontrun{
# stop AmCAT modules
stop_amcat_docker()

# stop container by id
stop_amcat_docker(filters = c(id = "a6cbe4787227"))

# stop all containers
stop_amcat_docker(filters = "")
}
}
