% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{upload_documents}
\alias{upload_documents}
\title{Upload documents}
\usage{
upload_documents(index, documents, columns = NULL, credentials = NULL)
}
\arguments{
\item{index}{The index name to create}

\item{documents}{A data frame with columns title, text, date, and optional other columns}

\item{columns}{An optional list with data types, e.g. list(author="keyword")}
}
\description{
Upload documents
}
